requirements.txt
playwright>=1.43
python-dotenv>=1.0
requests>=2.31
